// src/Core/enums.ts
export enum GTATunesEmoji {
    GTATUNES = '1429190161874550986',
    GTATUNES_BG = '1429865265964777555',
    PLAY = '1429193049908576260',
    PAUSE = '1429193024037982269',
    NEXT = '1429193048897617993',
    PREVIOUS = '1429193048566530139',
    NEXT_STATION = '1429193047542861956',
    PREVIOUS_STATION = '1429193046716580012',
    STOP = '1429193563488391228',
    GTA_III = '1429194851588505681',
    GTA_VC = '1429194855221035089',
    GTA_SA = '1429194853144858664',
    GTA_IV = '1429194852305731645',
    GTA_V = '1429194853992108122',
    SETTINGS = '1429855484357120091',
    DELETE_CONTROLLER = '1429253394224123954',
    EXTERNAL_LINK = '1429865115351650394',
    PROGRESS_PART = '1429867384646406184',
    PROGRESS_INDICATOR = '1429867383685779478',
    III_CHATTERBOX_ICON = '1432079880186757313',
    III_DOUBLE_CLEF_FM_ICON = '1432079881038069851',
    III_FLASHBACK_FM_ICON = '1432079881486733454',
    III_GAME_RADIO_FM_ICON = '1432079882711601352',
    III_HEAD_RADIO_ICON = '1432079883990863892',
    III_K_JAH_ICON = '1432079885089767554',
    III_LIPS_106_ICON = '1432079886452920504',
    III_MSX_FM_ICON = '1432079887824584836',
    III_RISE_FM_ICON = '1432079888826761462',
    IV_ELECTRO_CHOC_ICON = '1432079889628135555',
    IV_FUSION_FM_ICON = '1432079890634506445',
    IV_INDEPENDENCE_FM_ICON = '1432079891301404692',
    IV_INTEGRITY_2_0_ICON = '1432079892337397954',
    IV_INTERNATIONAL_FUNK_99_ICON = '1432079893193166928',
    IV_JAZZ_NATION_RADIO_ICON = '1432079894380282031',
    IV_K109_THE_STUDIO_ICON = '1432079895508553768',
    IV_LIBERTY_CITY_HARDCORE_ICON = '1432079896426844200',
    IV_LIBERTY_ROCK_ICON = '1432079897332940913',
    IV_MASSIVE_B_ICON = '1432079898314412112',
    IV_PUBLIC_LIBERTY_RADIO_ICON = '1432079899408994304',
    IV_RADIO_BROKER_ICON = '1432079900327805071',
    IV_RAMJAM_FM_ICON = '1432079901674180628',
    IV_SAN_JUAN_SOUNDS_ICON = '1432079902835867779',
    IV_SELF_ACTUALIZATION_FM_ICON = '1432079903951683634',
    IV_THE_BEAT_102_7_ICON = '1432079904895275069',
    IV_THE_CLASSICS_104_1_ICON = '1432079905868349540',
    IV_THE_JOURNEY_ICON = '1432079906593968198',
    IV_THE_VIBE_98_8_ICON = '1432079907575300178',
    IV_TUFF_GONG_ICON = '1432079908242456818',
    IV_VICE_CITY_FM_ICON = '1432079909597085696',
    IV_VLADIVOSTOK_FM_ICON = '1432079910633078996',
    IV_WKTT_RADIO_ICON = '1432079911492784249',
    SA_BOUNCE_FM_ICON = '1432079912365326450',
    SA_CSR_ICON = '1432079913195667558',
    SA_CUSTOM_STATIONS_ICON = '1432079914127065209',
    SA_K_DST_ICON = '1432079915070525550',
    SA_K_JAH_ICON = '1432079916052254891',
    SA_K_ROSE_ICON = '1432079917151031357',
    SA_MASTER_SOUNDS_ICON = '1432079918027505867',
    SA_PLAYBACK_FM_ICON = '1432079919466287296',
    SA_RADIO_LOS_SANTOS_ICON = '1432079920695214212',
    SA_RADIO_X_ICON = '1432079921777348639',
    SA_SFUR_ICON = '1432079922893033635',
    SA_WCTR_ICON = '1432079923681689744',
    VC_EMOTION_ICON = '1432079924696715395',
    VC_ESPANTOSO_ICON = '1432079925703086191',
    VC_FEVER_105_ICON = '1432079926776954920',
    VC_FLASH_FM_ICON = '1432079932607041598',
    VC_K_CHAT_ICON = '1432079934117117972',
    VC_MP3_PLAYER_ICON = '1432079935106711659',
    VC_VCPR_ICON = '1432079937673887826',
    VC_V_ROCK_ICON = '1432079936344031272',
    VC_WAVE_ICON = '1432079938508423179',
    VC_WILDSTYLE_ICON = '1432079939498283178'
}